statistics: move analyze context out
